Rename image.rs to image_raw.rs
