+++ embedded-graphics/src/image/image_raw.rs
